Rewrite logger.py
Cleanup mostly everything, encapsule everyting in a Doorman class
making it possible to use it as a library.

There are more opportunities for cleanup:
- optional parameters that are not null in db
- use python logging instead of homegrown file log and print
- recycle the cursor
- split check_user in smaller helpers
- use fixtures for db data

Still Fablab's profiles are hardcoded which sucks. Fixing that
requires fixing the db layout which i think is wrong: profiles
should be stored in the db and there should be only one table
for the users.